[fixes] patch fix to setup dropbox settings

+++ frappe/patches/v7_0/setup_integration_services.py
@@ -3,24 +3,29 @@ import frappe
 from frappe.exceptions import DataError
 from frappe.utils.password import get_decrypted_password
 
+app_list = [
+	{"app_name": "razorpay_integration", "service_name": "Razorpay", "doctype": "Razorpay Settings", "remove": True},
+	{"app_name": "paypal_integration", "service_name": "PayPal", "doctype": "PayPal Settings", "remove": True},
+	{"app_name": "frappe", "service_name": "Dropbox Integration", "doctype": "Dropbox Backup", "remove": False}
+]
+
 def execute():
 	if not frappe.db.exists("DocType", "Integration Service"):
 		return
-	
+
 	installed_apps = frappe.get_installed_apps()
-	print installed_apps
-	for app_details in [{"app_name": "razorpay_integration", "service_name": "Razorpay"},
-		{"app_name":"paypal_integration", "service_name": "PayPal"}]:
+
+	for app_details in app_list:
 		if app_details["app_name"] in installed_apps:
 			try:
 				setup_integration_service(app_details)
-				uninstall_app(app_details["app_name"])
+				if app_details["remove"]:
+					uninstall_app(app_details["app_name"])
 			except Exception:
 				pass
 
 def setup_integration_service(app_details):
-	settings = get_app_settings(app_details["service_name"])
-
+	settings = get_app_settings(app_details)
 	if not settings:
 		raise DataError
 	
@@ -36,29 +41,51 @@ def setup_integration_service(app_details):
 	integration_service.flags.ignore_mandatory = True
 	integration_service.save(ignore_permissions=True)
 
-def get_app_settings(service_name):
+def get_app_settings(app_details):
 	from frappe.integration_broker.doctype.integration_service.integration_service import get_integration_controller
 	
 	parameters = []
-	doctype = docname = "{0} Settings".format(service_name)
-	settings = frappe.get_doc(doctype)
-	controller = get_integration_controller(service_name, setup=False)
+	doctype = docname = app_details["doctype"]
 	
-	if not settings.get(controller.parameters_template[0]["fieldname"]) and service_name == "PayPal":
-		settings = { "api_username": frappe.conf.paypal_username, "api_password": frappe.conf.paypal_password,
-			"signature": frappe.conf.paypal_signature }
+	settings = get_parameters(app_details)
+	controller = get_integration_controller(app_details["service_name"], setup=False)
 	
 	for d in controller.parameters_template:
-		if not settings.get(d.fieldname):
-			raise DataError
+		if settings.get(d.fieldname):
+			if ''.join(set(settings.get(d.fieldname))) == '*':
+				setattr(settings, d.fieldname, get_decrypted_password(doctype, docname, d.fieldname, raise_exception=True))
 
-		if ''.join(set(settings.get(d.fieldname))) == '*':
-			setattr(settings, d.fieldname, get_decrypted_password(doctype, docname, d.fieldname, raise_exception=False))
+			parameters.append({'label': d.label, 'fieldname': d.fieldname, "value": settings.get(d.fieldname)})
 
-		parameters.append({'label': d.label, 'fieldname': d.fieldname, "value": settings.get(d.fieldname)})
-	
 	return parameters
 
 def uninstall_app(app_name):
 	from frappe.installer import remove_from_installed_apps
 	remove_from_installed_apps(app_name)
+
+def get_parameters(app_details):
+	if app_details["service_name"] == "Razorpay":
+		return frappe.get_doc(app_details["doctype"])
+
+	elif app_details["service_name"] == "PayPal":
+		if frappe.conf.paypal_username and frappe.conf.paypal_password and frappe.conf.paypal_signature:
+			return {
+				"api_username": frappe.conf.paypal_username,
+				"api_password": frappe.conf.paypal_password,
+				"signature": frappe.conf.paypal_signature
+			}
+		else:
+			return frappe.get_doc(app_details["doctype"])
+
+	elif app_details["service_name"] == "Dropbox Integration":
+		doc = frappe.get_doc(app_details["doctype"])
+
+		if not (frappe.conf.dropbox_access_key and frappe.conf.dropbox_secret_key):
+			raise DataError
+
+		return {
+			"app_access_key": frappe.conf.dropbox_access_key,
+			"app_secret_key": frappe.conf.dropbox_secret_key,
+			"dropbox_access_key": doc.dropbox_access_key,
+			"dropbox_access_secret": doc.dropbox_access_secret
+		}
